add force parameter

+++ rkstool/force_add_qpfile.py
@@ -1,9 +1,13 @@
 import os
-import glob
 import shutil
 
 
-def addqpfile(dir: str, qp_ext: str = '.24.qpfile', accept_ext = ['.m2ts']):
+def addqpfile(
+    dir: str, 
+    qp_ext: str = '.24.qpfile', 
+    force: bool = True,  # whether to cover existed '*.qpfile'
+    accept_ext = ['.m2ts']
+):
     dir = os.path.abspath(dir)
     for fn in os.listdir(dir):
         fp = os.path.join(dir, fn)
@@ -13,16 +17,17 @@ def addqpfile(dir: str, qp_ext: str = '.24.qpfile', accept_ext = ['.m2ts']):
         name, ext = os.path.splitext(fn)
         if ext not in accept_ext:
             continue
+        waste_dir = os.path.join(dir, 'waste_qp')
         tar_qp_fp = os.path.join(dir, name + qp_ext)
         res_qp_fp = os.path.join(dir, name + '.qpfile')
+        if os.path.exists(res_qp_fp) and not force:
+            continue
         if os.path.exists(tar_qp_fp):
             if tar_qp_fp != res_qp_fp:
+                if os.path.exists(res_qp_fp):
+                    os.makedirs(waste_dir, exist_ok=True)
+                    shutil.move(res_qp_fp, waste_dir)
                 os.rename(tar_qp_fp, res_qp_fp)
         else:
             with open(res_qp_fp, 'w') as qpf:
                 qpf.write('0 K')
-        query = os.path.join(dir, name + '*.qpfile')
-        for waste_qpfile in glob.glob(query):
-            waste_dir = os.path.join(dir, 'waste_qp')
-            os.makedirs(waste_dir, exist_ok=True)
-            shutil.move(waste_qpfile, waste_dir)
